Moved ExecutionMetrics definition from Execution Trace module to transaction_receipt.rs
